feat(suite): split up the CoinjoinStatusWheel component
